refactor: rename address model files and update imports to authentication package
